mobile_app/backend/pyproject.toml
[project]
name = "hualaowei-chatbot"
version = "0.1.0"
description = "HuaLaoWei Chatbot Service for municipal services in Singapore."
authors = [
    { name = "Woo Yan Seun", email = "wooyanseun@gmail.com" },
    { name = "Jerick Cheong", email = "jerickcheong@gmail.com" },
    { name = "Fleming Siow", email = "flemingsiow@gmail.com" },
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "requests>=2.31.0",
    "ffmpeg-python>=0.2.0",
    "python-jose>=3.3.0",
    "psycopg[binary,pool]>=3.1.0",
    "pydantic>=2.0.0",
    "langchain-core>=0.1.0",
    "langid>=1.1.0",
    "minio>=7.1.14",
    "piexif>=1.1.3",
    "pillow>=10.0.0",
    "difflib2>=0.1.3", # Note: standard difflib is in Python stdlib, optional alternative for advanced needs
    "weaviate-client>=4.5.4"
]

[build-system]
requires = ["hatchling>=1.5"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]
include = ["*.py"]
